requirements.txt
bentoml
torchvision
joblib
tqdm
pip-chill
wincertstore
boto3
-e .